Slight refinement in the how the FAPI character generation detects we are producing anti-aliased glyphs.
Also, it seems that when assigning variable value to a second, larger data
type, whilst gcc promotes the data type before the assignment, MSVC promotes
the variable later, so shifting a 32 bit integer up during an assignment to
a 64 bit integer variable works find on gcc, but MSVC loses the most
significant bits. So we need to force the promotion to happen with a cast.

Bug 691588

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11695 a1074d23-0009-0410-80fe-cf8c14f379e6

+++ gs/psi/fapi_ft.c
@@ -1104,8 +1104,8 @@ static int move_to(const FT_Vector *aTo, void *aObject)
      * be able to convert to GS co-ordinates properly.
      */
      /* FAPI now expects these coordinates in 32.32 */
-    p->x = aTo->x << 26;
-    p->y = aTo->y << 26;
+    p->x = ((int64_t)aTo->x) << 26;
+    p->y = ((int64_t)aTo->y) << 26;
 
     return p->path->moveto(p->path, p->x, p->y) ? -1 : 0;
 }
@@ -1115,8 +1115,8 @@ static int line_to(const FT_Vector *aTo, void *aObject)
     FF_path_info *p = (FF_path_info*)aObject;
 
     /* See move_to() above */
-    p->x = aTo->x << 26;
-    p->y = aTo->y << 26;
+    p->x = ((int64_t)aTo->x) << 26;
+    p->y = ((int64_t)aTo->y) << 26;
 
     return p->path->lineto(p->path, p->x, p->y) ? -1 : 0;
 }
@@ -1169,10 +1169,10 @@ static int cubic_to(const FT_Vector *aControl1, const FT_Vector *aControl2, cons
     p->x = aTo->x << 26;
     p->y = aTo->y << 26;
 
-    Control1x = aControl1->x << 26;
-    Control1y = aControl1->y << 26;
-    Control2x = aControl2->x << 26;
-    Control2y = aControl2->y << 26;
+    Control1x = ((int64_t)aControl1->x) << 26;
+    Control1y = ((int64_t)aControl1->y) << 26;
+    Control2x = ((int64_t)aControl2->x) << 26;
+    Control2y = ((int64_t)aControl2->y) << 26;
     return p->path->curveto(p->path, Control1x, Control1y, Control2x, Control2y, p->x, p->y) ? -1 : 0;
 
     p->x = aTo->x;
